fix: distribute functions between two utils files and add funcs to header
